apps/consumers/vital-timeseries/requirements.txt
kafka-python==2.0.2
influxdb-client==1.48.0
pandas==1.5.3
